requirements.txt
# TODO: put back the standard pypi release when
# next point version of openwisp-controller is released
openwisp-controller @ https://github.com/openwisp/openwisp-controller/tarball/master
# openwisp-utils[rest] is left unpinned so the version used
# is the one dictated by openwisp-controller
influxdb~=5.3
django-celery-email~=3.0.0
mac-vendor-lookup~=0.1
django-cache-memoize~=0.1
django-nested-admin~=3.3.2
swapper~=1.1
netaddr~=0.8
python-dateutil>=2.7,<3.0
